+++ src/Game/Werewolf/Engine.hs
@@ -44,7 +44,8 @@ module Game.Werewolf.Engine (
     createPlayers,
 
     -- ** Queries
-    doesPlayerExist, isPlayerDefender, isPlayerSeer, isPlayerWildChild, isPlayerWitch,
+    doesPlayerExist,
+    isPlayerDefender, isPlayerSeer, isPlayerVillageIdiot, isPlayerWildChild, isPlayerWitch,
     isPlayerWolfHound,
     isPlayerWerewolf,
     isPlayerAlive, isPlayerDead,
@@ -138,16 +139,19 @@ checkStage' = use stage >>= \stage' -> case stage' of
         advanceStage
 
     VillagesTurn -> do
-        playersCount    <- uses players (length . filterAlive)
+        alivePlayers    <- uses players filterAlive
+        playersCount    <- ifM (use villageIdiotRevealed)
+            (return . length $ filter (not . isVillageIdiot) alivePlayers)
+            (return $ length alivePlayers)
         votes'          <- use votes
 
         when (playersCount == Map.size votes') $ do
             tell $ map (uncurry playerMadeLynchVoteMessage) (Map.toList votes')
 
             getVoteResult >>= \votees -> case votees of
-                [votee]   -> do
-                    killPlayer $ votee ^. name
-                    tell [playerLynchedMessage votee]
+                [votee]   -> if isVillageIdiot votee
+                    then villageIdiotRevealed .= True >> tell [villageIdiotLynchedMessage $ votee ^. name]
+                    else killPlayer (votee ^. name) >> tell [playerLynchedMessage votee]
                 _               ->
                     findPlayerByRole scapegoatRole >>= \mScapegoat -> case mScapegoat of
                         Just scapegoat  -> killPlayer (scapegoat ^. name) >> tell [scapegoatLynchedMessage (scapegoat ^. name)]
@@ -349,6 +353,9 @@ isPlayerDefender name = isDefender <$> findPlayerByName_ name
 
 isPlayerSeer :: MonadState Game m => Text -> m Bool
 isPlayerSeer name = isSeer <$> findPlayerByName_ name
+
+isPlayerVillageIdiot :: MonadState Game m => Text -> m Bool
+isPlayerVillageIdiot name = isVillageIdiot <$> findPlayerByName_ name
 
 isPlayerWildChild :: MonadState Game m => Text -> m Bool
 isPlayerWildChild name = isWildChild <$> findPlayerByName_ name